refactor(adapters): extract shared worker/reviewer result parsing
gpt, local, and openrouter each carried a byte-identical (modulo formatting)
copy of eight result-parsing functions, already drifting in comments — a parse
fix would have needed three-way sync. New src/adapters/resultParsing.ts exports
parseWorkerResult / parseReviewerResult; the three adapters delegate in one line
and drop their local copies (and the now-orphaned t import).

codex is intentionally left on its own variant: its text fallbacks salvage
filesChanged/commands/issues from prose and its JSON extractor differs, so
sharing would regress behavior or touch untested paths.

Adapter suite 76/76 green. The openrouter.ts edit also carries the INT-1460
key-source change.

Refs: INT-1441
